Implement `verifyProperty` helper function
The tests for specific property descriptors may have side effects on the
object under test. This imposes implicit invocation order restrictions
if they are to be used together on the same object.

Expose a higher-order function for testing multiple property descriptors
that, while destructive, avoids inconsistent behavior that could result
from using the specific tests in an invalid sequence.