Add tests for Annex B.3.4
B.3.4 FunctionDeclarations in IfStatement Statement Clauses

The following rules for IfStatement augment those in 13.6:

    IfStatement[Yield, Return]:
        if ( Expression[In, ?Yield] ) FunctionDeclaration[?Yield] else Statement[?Yield, ?Return]
        if ( Expression[In, ?Yield] ) Statement[?Yield, ?Return] else FunctionDeclaration[?Yield]
        if ( Expression[In, ?Yield] ) FunctionDeclaration[?Yield] else FunctionDeclaration[?Yield]
        if ( Expression[In, ?Yield] ) FunctionDeclaration[?Yield]

The above rules are only applied when parsing code that is not strict
mode code. If any such code is match by one of these rules subsequent
processing of that code takes places as if each matching occurrence of
FunctionDeclaration[?Yield] was the sole StatementListItem of a
BlockStatement occupying that position in the source code. The semantics
of such a synthetic BlockStatement includes the web legacy compatibility
semantics specified in B.3.3.