lint: Define harness features in harness source
Previously, the `HarnessFeatures` linting check verified that tests
which included certain harness files also explicitly declared the
language features used by those files. This relied on a dedicated file
to declare which harness files used which features. Because the harness
files support the inclusion of metadata, this indirection was not
strictly necessary.

Relocate the "features" declarations into the harness files themselves.
This makes the pattern easier for others to discovers and more closely
aligns with how the metadata is defined in the test files. Refactor the
`Features` linting check to perform this validation.

Correct a bug in the original implementation, and fix the issues which
the corrected linting check identifies.