Remove assert.deepEqual
The `assert.deepEqual` utility method was originally implemented in
support of two tests for the "RegExp Match Indices" proposal [1]. In the
years since that time, it has been referenced by one additional test,
where it is interchangeable with the considerably simpler
`assert.compareArray`.

Refactor the RegExp tests to use a simpler inline helper function that
verifies the same semantics and preserves meaningful error messages.

[1] https://github.com/tc39/test262/pull/2309