Validate WEB_FEATURES.yml
For every pull request, verify that the `WEB_FEATURES.yml` file is
internally consistent and that the pull request does not reduce the
number of tests associated with a given classifier.

Failing for any reduction in the number of matched tests is intended to
alert contributors during file renaming operations. This failure will
need to be ignored when tests are simply deleted, and this could
potentially become a distraction for maintainers. However, test deletion
has historically been so rare that spurious failures are not expected to
be common enough to substantively interfere with routine maintenance.